rewrite join logic to ideally allow for 3 attempts before a kick happens
